requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-cors==4.0.1
flask-jwt-extended==4.6.0
celery==5.4.0
pandas==2.2.2